Restructured startup service configuration code
