+++ src/Dash/src/Dash/Application/Default/Startup.cs
@@ -1,6 +1,9 @@
 ﻿// Copyright (c) Huy Hoang. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
 
+using Dash.Common;
+using Dash.Engine.Repositories;
+using Dash.Engine.Visitors;
 using Dash.PreprocessingSteps;
 using Dash.Roslyn;
 using Microsoft.Extensions.DependencyInjection;
@@ -14,15 +17,16 @@ namespace Dash.Application.Default
             var services = new ServiceCollection();
 
             ApplicationServices.Add(services, dashOptions);
-            CommonServices.Add(services);
-            RepositoryServices.Add(services);
             SourceCodeServices.Add(services);
-            NodeVisitorServices.Add(services);
             LanguageProviderServices.Add(services);
             GeneratorServices.Add(services);
 
+            services.AddHttpClient();
+            services.AddCommonServices();
             services.AddPreprocessingSteps();
             services.AddRoslynFacade();
+            services.AddRepositories();
+            services.AddNodeVisitors();
 
             return services;
         }
